DDF-1418: Moves geocoder modules from ui to spatial
