+++ catalog/ui/search-ui/pom.xml
@@ -35,9 +35,6 @@
         <module>search-redirect</module>
         <module>search-htmltransformer</module>
         <module>search-endpoint</module>
-        <module>search-geocoder</module>
-        <module>search-geonames</module>
-        <module>search-geonames-local</module>
         <module>search-ui-app</module>
     </modules>
