--- catalog/ui/search-ui/search-geocoder/src/main/java/org/codice/ddf/ui/searchui/geocoder/GeoResult.java
@@ -1,52 +1,0 @@
-/**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser
- * General Public License as published by the Free Software Foundation, either version 3 of the
- * License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
- * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details. A copy of the GNU Lesser General Public License
- * is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/
-package org.codice.ddf.ui.searchui.geocoder;
-
-import java.util.List;
-
-import org.opengis.geometry.primitive.Point;
-
-public class GeoResult {
-
-    protected Point point;
-
-    protected List<Point> bbox;
-
-    protected String fullName;
-
-    public Point getPoint() {
-        return point;
-    }
-
-    public void setPoint(Point point) {
-        this.point = point;
-    }
-
-    public List<Point> getBbox() {
-        return bbox;
-    }
-
-    public void setBbox(List<Point> bbox) {
-        this.bbox = bbox;
-    }
-
-    public String getFullName() {
-        return fullName;
-    }
-
-    public void setFullName(String fullName) {
-        this.fullName = fullName;
-    }
-}
